Add: Proper error handling

+++ backend/controllers/product.controller.js
@@ -15,7 +15,7 @@ export const createProduct = async (req, res) => {
     const product = req.body; //user will send product data in the request body
 
     if(!product.name || !product.price || !product.image) {
-        return res.status(400).json({ success:false, message: 'Please provide all the fields' });
+        return res.status(404).json({ success:false, message: 'Please provide all the fields' });
     }
 
     const newProduct = new Product(product)
@@ -36,7 +36,7 @@ export const updateProduct = async (req, res) => {
     const product = req.body;
 
     if(!mongoose.Types.ObjectId.isValid(id)) {
-        return res.status(400).json({ success: false, message: 'Invalid product ID' });
+        return res.status(404).json({ success: false, message: 'Invalid product ID' });
     }
 
     try {
@@ -50,11 +50,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: 'Invalid product ID' });
+    }
+
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: 'Product deleted successfully'});
     } catch (error) {
         console.log("Error deleting product:", error.message);
-        res.status(404).json({ success: false, message: 'Product not found' });
+        res.status(500).json({ success: false, message: 'Server Error' });
     }
 }
